PenaltyManager service

+++ Library/Member.cs
@@ -4,9 +4,9 @@
     {
         public int MemberId { get; set; }
         public bool IsMemberStudent { get; set; }
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
 
-        public Member (int MemberId, bool IsMemberStudent, double balance)
+        public Member (int MemberId, bool IsMemberStudent, decimal balance)
         {
             this.MemberId = MemberId;
             this.IsMemberStudent = IsMemberStudent;
